[#D3D-6202] Updates Sketchfab plugin code to match updated import
This commit updates the Sketchfab-side code and patch to take into
account changes to the GlTF import functions.

It also switches the submodule repository to Khronos' one instead
of a previous fork of their repo which lived in the Sketchfab
Github organization.
Updates to commit 709630548cdc184af6ea50b2ff3ddc5450bc0af3

An uniformization of licenses (GPL to Apache) is also done.

Note: no change in behaviour should happen.